# openapi/crm-mep.yaml
openapi: 3.1.0
...
info:
  title: CRM Frisson ↔ MEP-LEAN
  version: 0.2.0-draft
  summary: Seis operaciones del CRM Frisson como servidor de la integración con MEP-LEAN.
  description: |
    El CRM Frisson conserva la autoridad comercial (P-01) y expone toda la
    superficie física del contrato (P-02, pull + write-back). MEP-LEAN es
    cliente: consulta el intake, lee contexto de OUV, acusa recepción técnica
    y publica respuestas comerciales versionadas.

    Transporte: HTTPS obligatorio (TLS 1.2+). HTTP → 426.
    Fechas: UTC, RFC 3339, sufijo `Z`.
    ETags: fuertes, opacos y entrecomillados (nunca `W/`).
servers:
  - url: https://sandbox.api.frisson.crm
    description: Sandbox (cuotas = 1/4 de producción)
  - url: https://staging.api.frisson.crm
    description: Staging
  - url: https://api.frisson.crm
    description: Producción
security:
  - apiKeyAuth: []
...
paths:
...
  /v1/commercial-opportunities/{opportunity_ref}:
    get:
      tags: [Opportunity]
      operationId: getCommercialOpportunity
      summary: Contexto puntual de oportunidad (OUV)
      description: |
        INV-09: los nulos se preservan. Un campo sin valor se serializa como
        `null`; nunca se omite ni se sustituye por `""`, `0` o un placeholder.
        INV-11: `commercial_archetype` es autoridad del CRM. MEP puede usarlo
        para su proyección operativa interna, pero no lo devuelve ni lo
        sobrescribe por este contrato, y no se equipara automáticamente con el
        carril operativo `Archetype_Lane`.

        Solo lectura: no existe ninguna operación de escritura sobre la OUV.
      security:
        - apiKeyAuth: [opportunities:read]
      parameters:
        - $ref: '#/components/parameters/OpportunityRef'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Contexto de la OUV
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommercialOpportunity' }
              examples:
                ouv-del-brief:
                  $ref: '#/components/examples/OpportunityFull'
                ouv-con-nulos:
                  $ref: '#/components/examples/OpportunityNulls'
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/ouv_context/{ouv_id}:
    get:
      tags: [Opportunity]
      operationId: getOuvContext
      summary: Lectura contextual de la OUV y su historial con MEP-LEAN
      description: |
        `{ouv_id}` es el consecutivo de la OUV (`OUV-####`), el mismo
        identificador que `crm_opportunity_ref`.

        `opportunity` es la misma proyección que
        `GET /v1/commercial-opportunities/{opportunity_ref}`. Si la OUV aún no
        tiene fila espejo, se proyecta desde `ouvs` sin crearla
        (`source_version` = `0`).

        `mep_interactions` es el historial que la UI muestra en Solicitudes de
        Preventa: la solicitud del CRM (`commercial_interaction`) y, en cada
        una, `narrativa` (respuestas MEP, de la más reciente a la más antigua)
        y `pista_tecnica` (acuses). No incluye `semantic_fingerprint` ni
        `payload_hash`.

        `mep_interactions.items` es siempre el listado completo de solicitudes
        de esa OUV, de la más reciente a la más antigua. Sin solicitudes, solo
        trae `items: []`. Solo lectura: no actualiza la OUV, su versión ni su
        ETag persistido. El header `ETag` es del agregado de esta respuesta.
      security:
        - apiKeyAuth: [opportunities:read]
      parameters:
        - $ref: '#/components/parameters/OuvContextId'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Oportunidad y listado completo de interacciones con MEP-LEAN
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OuvContext' }
              examples:
                con-historial:
                  $ref: '#/components/examples/OuvContextWithHistory'
                sin-solicitudes:
                  $ref: '#/components/examples/OuvContextEmpty'
        '304': { $ref: '#/components/responses/NotModified' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Credencial no humana, por ambiente, fuera de payloads, logs y Git.
        Admite rotación y revocación. 401 = ausente o inválida;
        403 = identidad válida sin permiso para la operación solicitada.
  parameters:
...
    OpportunityRef:
      name: opportunity_ref
      in: path
      required: true
      schema: { type: string, maxLength: 64 }
...
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema: { type: string, maxLength: 128 }
...
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      schema: { type: string }
    OuvContextId:
      name: ouv_id
      in: path
      required: true
      description: Consecutivo de la OUV (`OUV-####`).
      schema: { type: string, maxLength: 64 }
  headers:
    ETag:
      description: Versión opaca fuerte y entrecomillada del recurso.
      schema: { type: string }
    XCorrelationId:
      description: Eco del `X-Correlation-ID` recibido, o uno generado.
      schema: { type: string }
...
    RateLimitRemaining:
      schema: { type: integer }
...
    RetryAfter:
      description: Segundos hasta poder reintentar.
      schema: { type: integer }
  schemas:
    ServiceHorizon:
      type: string
      enum: [IMMEDIATE, DEFERRED, UNSPECIFIED]
    ServiceName:
      type: string
      enum: [TECHNICAL_DESIGN, FINANCIAL_DESIGN]
    ServiceDependency:
      type: string
      enum: [NONE, TECHNICAL_DESIGN, FINANCIAL_DESIGN]
    ProcessingStatus:
      type: string
      enum: [ACCEPTED, DUPLICATE, QUARANTINED, REJECTED]
    BusinessMilestone:
      type: string
      enum:
        - INTERACTION_RECEIVED
        - ENGINEER_ASSIGNED
        - ROUTE_CAPACITY_REGISTERED
        - INTERACTION_COMPLETED
    ResponseStatus:
      type: string
      enum: [RECEIVED, IN_PROGRESS, PARTIALLY_COMPLETED, COMPLETED]
    ServiceResultStatus:
      type: string
      enum: [RECEIVED, IN_PROGRESS, COMPLETED, CANCELLED]
    ServiceOutcome:
      description: |
        Enum de outcome del servicio. `CONDITIONED` permite representar
        entregas hechas bajo una condición externa acordada, distinta de
        una entrega parcial no acabada (R2 §6.4).
      type: [string, 'null']
      enum: [VIABLE, NOT_VIABLE, PARTIAL, CONDITIONED, null]
    RouteStatus:
      type: string
      enum: [VIABLE, NOT_VIABLE, CONDITIONED]
    CapacityStatus:
      type: string
      enum: [PLANNED, NOT_PLANNED, CONDITIONED]
    OpportunityStatus:
      description: Espejo de `ouvs.resultado` en el CRM (autoridad del CRM).
      type: [string, 'null']
      enum: [EnCurso, Ganada, Perdida, Descartada, null]
...
    CommercialOpportunity:
      type: object
      additionalProperties: false
      required:
        - crm_opportunity_ref
        - title
        - organization
        - commercial_value
        - stage
        - status
        - expected_close_date
        - commercial_owner
        - commercial_archetype
        - context_observed_at
        - source_version
        - etag
      properties:
        crm_opportunity_ref: { type: string, maxLength: 64 }
        title: { type: [string, 'null'], maxLength: 512 }
        organization:
          type: [object, 'null']
          additionalProperties: false
          required: [ref, name]
          properties:
            ref: { type: string, maxLength: 64 }
            name: { type: [string, 'null'], maxLength: 512 }
        commercial_value:
          type: [object, 'null']
          additionalProperties: false
          required: [amount, currency]
          properties:
            amount:
              description: Entero en la unidad menor no fraccionada (COP sin decimales).
              type: integer
            currency: { type: [string, 'null'], minLength: 3, maxLength: 3 }
        stage:
          type: [object, 'null']
          additionalProperties: false
          required: [ref, name]
          properties:
            ref: { type: string, maxLength: 64 }
            name: { type: [string, 'null'], maxLength: 256 }
        status: { $ref: '#/components/schemas/OpportunityStatus' }
        expected_close_date: { type: [string, 'null'], format: date }
        commercial_owner:
          type: [object, 'null']
          additionalProperties: false
          required: [ref, display_name]
          properties:
            ref: { type: string, maxLength: 64 }
            display_name: { type: [string, 'null'], maxLength: 256 }
        commercial_archetype:
          description: Autoridad CRM. Referencia y nombre estables o `null` (INV-10/INV-11).
          type: [object, 'null']
          additionalProperties: false
          required: [ref, name]
          properties:
            ref: { type: string, maxLength: 64 }
            name: { type: [string, 'null'], maxLength: 256 }
        context_observed_at: { type: string, format: date-time }
        source_version: { type: string, maxLength: 32 }
        etag: { type: string, maxLength: 96 }
    OuvContext:
      type: object
      additionalProperties: false
      required: [etag, opportunity, mep_interactions]
      properties:
        etag:
          description: |
            ETag del agregado. Idéntico al header `ETag`. No se persiste y
            no modifica el `etag` de la OUV.
          type: string
        opportunity: { $ref: '#/components/schemas/CommercialOpportunity' }
        mep_interactions:
          type: object
          additionalProperties: false
          required: [items]
          properties:
            items:
              description: Listado completo de solicitudes de la OUV.
              type: array
              items: { $ref: '#/components/schemas/OuvContextInteraction' }
    OuvContextInteraction:
      description: |
        Solicitud CRM → MEP-LEAN y su historial, la misma proyección que ve
        la UI de Preventa. Sin `semantic_fingerprint` ni `payload_hash`.
      type: object
      additionalProperties: false
      required:
        - crm_interaction_ref
        - crm_opportunity_ref
        - subject
        - service_horizon
        - requested_services
        - source_content
        - sharepoint_document_url
        - source_created_at
        - source_version
        - etag
        - estado
        - asignacion
        - ruta_capacidad
        - servicios
        - planner_url
        - clasificacion_entregada
        - narrativa
        - pista_tecnica
        - polling_status
      properties:
        crm_interaction_ref: { type: string, maxLength: 64 }
        crm_opportunity_ref: { type: [string, 'null'], maxLength: 64 }
        subject: { type: [string, 'null'], maxLength: 512 }
        service_horizon: { $ref: '#/components/schemas/ServiceHorizon' }
        requested_services:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [service, label, dependency]
            properties:
              service: { $ref: '#/components/schemas/ServiceName' }
              label: { type: string }
              dependency: { $ref: '#/components/schemas/ServiceDependency' }
        source_content: { type: string }
        sharepoint_document_url: { type: [string, 'null'] }
        source_created_at: { type: [string, 'null'], format: date-time }
        source_version: { type: string }
        etag: { type: string }
        estado:
          type: object
          additionalProperties: false
          required: [hito, response_status, eta_date, next_milestone, sin_respuesta_mep]
          properties:
            hito:
              oneOf:
                - $ref: '#/components/schemas/BusinessMilestone'
                - type: 'null'
            response_status:
              oneOf:
                - $ref: '#/components/schemas/ResponseStatus'
                - type: 'null'
            eta_date: { type: [string, 'null'], format: date }
            next_milestone: { type: [string, 'null'] }
            sin_respuesta_mep: { type: boolean }
        asignacion:
          type: [object, 'null']
          additionalProperties: false
          required: [engineer, assigned_at]
          properties:
            engineer:
              type: object
              additionalProperties: false
              required: [ref, display_name]
              properties:
                ref: { type: string }
                display_name: { type: string }
            assigned_at: { type: [string, 'null'], format: date-time }
        ruta_capacidad:
          type: [object, 'null']
          additionalProperties: false
          required: [version, route_status, capacity_status, summary, registered_at, registro_url]
          properties:
            version: { type: string }
            route_status:
              oneOf:
                - $ref: '#/components/schemas/RouteStatus'
                - type: 'null'
            capacity_status:
              oneOf:
                - $ref: '#/components/schemas/CapacityStatus'
                - type: 'null'
            summary: { type: [string, 'null'] }
            registered_at: { type: [string, 'null'], format: date-time }
            registro_url: { type: [string, 'null'] }
        servicios:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - service
              - label
              - status
              - outcome
              - dependency
              - summary
              - reason_code
              - bloqueado_por_dependencia
              - entregables
            properties:
              service: { $ref: '#/components/schemas/ServiceName' }
              label: { type: string }
              status: { $ref: '#/components/schemas/ServiceResultStatus' }
              outcome: { $ref: '#/components/schemas/ServiceOutcome' }
              dependency: { $ref: '#/components/schemas/ServiceDependency' }
              summary: { type: [string, 'null'] }
              reason_code: { type: [string, 'null'] }
              bloqueado_por_dependencia: { type: boolean }
              entregables:
                type: array
                items:
                  type: object
                  additionalProperties: false
                  required: [url, label, published_at]
                  properties:
                    url: { type: string }
                    label: { type: [string, 'null'] }
                    published_at: { type: [string, 'null'], format: date-time }
        planner_url: { type: [string, 'null'] }
        clasificacion_entregada: { type: [string, 'null'] }
        narrativa:
          description: Respuestas MEP, de la más reciente a la más antigua.
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - response_version
              - business_milestone
              - response_status
              - narrative_note
              - responded_at
              - responded_by
            properties:
              response_version: { type: integer, minimum: 1 }
              business_milestone: { $ref: '#/components/schemas/BusinessMilestone' }
              response_status: { $ref: '#/components/schemas/ResponseStatus' }
              narrative_note: { type: [string, 'null'] }
              responded_at: { type: [string, 'null'], format: date-time }
              responded_by:
                type: object
                additionalProperties: false
                required: [ref, display_name]
                properties:
                  ref: { type: string }
                  display_name: { type: string }
        pista_tecnica:
          description: Acuses técnicos, separados de la narrativa.
          type: array
          items:
            type: object
            additionalProperties: false
            required: [receipt_id, receipt_version, processing_status, reason_code, observed_at]
            properties:
              receipt_id: { type: string }
              receipt_version: { type: integer, minimum: 0 }
              processing_status: { $ref: '#/components/schemas/ProcessingStatus' }
              reason_code: { type: [string, 'null'] }
              observed_at: { type: [string, 'null'], format: date-time }
        polling_status:
          oneOf:
            - $ref: '#/components/schemas/ProcessingStatus'
            - type: 'null'
...
    Problem:
      description: RFC 7807. Nunca incluye `source_content`, `X-API-Key` ni IDs internos de MEP (INV-02).
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        correlation_id: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [pointer, code]
            properties:
              pointer: { type: string }
              code: { type: string }
  responses:
    NotModified:
      description: Sin cambios respecto del `If-None-Match` enviado.
      headers:
        ETag: { $ref: '#/components/headers/ETag' }
    BadRequest:
      description: 'ERR-400 — solicitud mal formada.'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: 'ERR-401 — clave ausente, inválida, revocada o expirada. Cuerpo genérico.'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: 'ERR-403 — identidad válida sin el scope requerido.'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: 'ERR-404 — recurso inexistente o no visible para la identidad.'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    TooManyRequests:
      description: 'ERR-429 — cuota temporal excedida.'
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ServiceUnavailable:
      description: 'ERR-503 — error transitorio del CRM.'
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  examples:
...
    OpportunityFull:
      summary: OUV completa (ouv_9101)
      externalValue: ../backend/test/fixtures/opportunity/ouv-9101-full.json
    OpportunityNulls:
      summary: OUV con todos los opcionales en null
      externalValue: ../backend/test/fixtures/opportunity/ouv-9102-nulls.json
    OuvContextWithHistory:
      summary: OUV con una solicitud y la respuesta de MEP-LEAN
      value:
        etag: '"ouv-context-OUV-0001-9f3a1c0e"'
        opportunity:
          crm_opportunity_ref: OUV-0001
          title: Enlace norte
          organization: { ref: acc_1, name: Alcaldía }
          commercial_value: { amount: 1000000, currency: COP }
          stage: { ref: UNIVERSO, name: UNIVERSO }
          status: EnCurso
          expected_close_date: '2026-07-15'
          commercial_owner: { ref: user_1, display_name: Carlos Ruiz }
          commercial_archetype: { ref: seg_1, name: Gobierno central }
          context_observed_at: '2026-09-28T15:00:00.000Z'
          source_version: '4'
          etag: '"ouv-OUV-0001-4-EnCurso-6-2026-09-01T00:00:00.000Z"'
        mep_interactions:
          items:
            - crm_interaction_ref: int_ouv0001_1
              crm_opportunity_ref: OUV-0001
              subject: Diseño técnico
              service_horizon: IMMEDIATE
              requested_services:
                - service: TECHNICAL_DESIGN
                  label: Técnica
                  dependency: NONE
              source_content: Necesitamos viabilidad.
              sharepoint_document_url: https://contoso.sharepoint.com/sites/docs/a.docx
              source_created_at: '2026-09-01T10:00:00.000Z'
              source_version: '1'
              etag: '"int_ouv0001_1-v1"'
              estado:
                hito: ENGINEER_ASSIGNED
                response_status: IN_PROGRESS
                eta_date: '2026-09-20'
                next_milestone: null
                sin_respuesta_mep: false
              asignacion:
                engineer: { ref: eng_1, display_name: Ana Pérez }
                assigned_at: '2026-09-01T12:00:00.000Z'
              ruta_capacidad: null
              servicios:
                - service: TECHNICAL_DESIGN
                  label: Técnica
                  status: IN_PROGRESS
                  outcome: null
                  dependency: NONE
                  summary: En curso
                  reason_code: null
                  bloqueado_por_dependencia: false
                  entregables: []
              planner_url: null
              clasificacion_entregada: null
              narrativa:
                - response_version: 2
                  business_milestone: ENGINEER_ASSIGNED
                  response_status: IN_PROGRESS
                  narrative_note: Ingeniero asignado.
                  responded_at: '2026-09-01T12:00:00.000Z'
                  responded_by: { ref: mep_user, display_name: Preventa MEP }
              pista_tecnica:
                - receipt_id: rcpt_1
                  receipt_version: 1
                  processing_status: ACCEPTED
                  reason_code: null
                  observed_at: '2026-09-01T11:00:00.000Z'
              polling_status: ACCEPTED
    OuvContextEmpty:
      summary: OUV sin solicitudes de preventa
      value:
        etag: '"ouv-context-OUV-0001-empty"'
        opportunity:
          crm_opportunity_ref: OUV-0001
          title: Enlace norte
          organization: { ref: acc_1, name: Alcaldía }
          commercial_value: { amount: 1000000, currency: COP }
          stage: { ref: UNIVERSO, name: UNIVERSO }
          status: EnCurso
          expected_close_date: '2026-07-15'
          commercial_owner: { ref: user_1, display_name: Carlos Ruiz }
          commercial_archetype: { ref: seg_1, name: Gobierno central }
          context_observed_at: '2026-09-28T15:00:00.000Z'
          source_version: '0'
          etag: '"ouv-OUV-0001-v0"'
        mep_interactions:
          items: []
...